requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
autopep8==2.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap4==2024.1
Django==4.2
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-js-asset==2.2.0
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
openai==0.28.0
pillow==10.3.0
psycopg2==2.9.9
pycodestyle==2.11.1
PyJWT==2.8.0
redis==5.0.4
requests==2.32.2
sqlparse==0.5.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
